requirements.txt
streamlit
openai>=1.0.0
requests
fpdf
python-dotenv
beautifulsoup4
lxml